connection to bakcend

+++ route/userRoute.js
@@ -1,5 +1,5 @@
 const express = require('express')
-const { register, verifyEmail, resendVerification, forgetPassword, resetPassword, getAllUsers, getUserDetail, findUser, updateUser, removeUser, signIn, signOut, deleteAll } = require('../controller/userController')
+const { register, verifyEmail, resendVerification, forgetPassword, resetPassword, getAllUsers, getUserDetail, findUser, updateUser, removeUser, signIn, signOut, deleteAll, updateRoleAdmin, updateRoleUser } = require('../controller/userController')
 const { userCheck, validate } = require('../Validation')
 const router = express.Router()
 
@@ -12,10 +12,14 @@ router.get('/userslist',getAllUsers)
 router.get('/userdetail/:id',getUserDetail)
 router.get('/findusers', findUser)
 router.put('/updateuser/:id',updateUser)
-router.delete('/removeuser', removeUser)
+router.put('/updaterole/:id', updateRoleAdmin )
+router.put('/updateroleuser/:id', updateRoleUser )
+router.delete('/removeuser/:id', removeUser)
 router.delete('/deleteall', deleteAll)
 
 router.post('/signin', validate, userCheck, signIn)
 router.get('/signout', signOut)
+
+router.post('/adminrole', )
 
 module.exports = router
